Adding fixes to all errors except for EngineState being an unresolved external

+++ Dilgine/Dilgine/Scripts/Core/Vector2.h
@@ -4,11 +4,11 @@
 
 struct Vector2
 {
-	float x, y;
+	int x, y;
 
 	Vector2() { x = 0; y = 0; }
-	Vector2(float val) { x = val; y = val; }
-	Vector2(float valX, float valY) { x = valX; y = valY; }
+	Vector2(int val) { x = val; y = val; }
+	Vector2(int valX, int valY) { x = valX; y = valY; }
 
 	float Magnitude() { return sqrt((x * x) + (y * y)); }
 	Vector2 Normalized() {
@@ -20,10 +20,10 @@ struct Vector2
 
 	Vector2 operator+(const Vector2& rhs) { return Vector2(x + rhs.x, y + rhs.y); }
 	Vector2 operator-(const Vector2& rhs) { return Vector2(x - rhs.x, y - rhs.y); }
-	Vector2 operator*(const float rhs) { return Vector2(x * rhs, y * rhs); }
+	Vector2 operator*(const int rhs) { return Vector2(x * rhs, y * rhs); }
 	void operator+=(const Vector2& rhs) { x += rhs.x; y += rhs.y; }
 	void operator-=(const Vector2& rhs) { x -= rhs.x; y -= rhs.y; }
-	void operator*=(const float rhs) { x *= rhs; y *= rhs; }
+	void operator*=(const int rhs) { x *= rhs; y *= rhs; }
 };
 
 #endif
