+++ Dilgine/Dilgine/Scripts/Game/Components/RectangleRenderer.h
@@ -6,11 +6,11 @@
 class RectangleRenderer
 {
 public:
-	float width, height;
+	int width, height;
 	Vector3 color;
 
 	RectangleRenderer() { width = 1; height = 1; color = Vector3(255); }
-	RectangleRenderer(float vWidth, float vHeight, Vector3 vColor) { width = vWidth; height = vHeight; color = vColor; }
+	RectangleRenderer(int vWidth, int vHeight, Vector3 vColor) { width = vWidth; height = vHeight; color = vColor; }
 };
 
 #endif
